feat: create separate score components for PDF actions liees
