+++ app.territoiresentransitions.react/src/app/pages/collectivite/PlansActions/ExportPdf/FicheActionPdf/actions-liees/ActionsLiees.tsx
@@ -1,10 +1,9 @@
-import { avancementToLabel, referentielToName } from '@/app/app/labels';
-import { actionAvancementColors } from '@/app/app/theme';
-import { ActionWithStatut } from '@/app/referentiels/actions/use-list-actions';
+import { referentielToName } from '@/app/app/labels';
+import { ScoreProgressBar } from '@/app/app/pages/collectivite/PlansActions/ExportPdf/scores/score-progress-bar';
+import { ScoreRatioBadge } from '@/app/app/pages/collectivite/PlansActions/ExportPdf/scores/score-ratio-badge';
+
 import {
-  Badge,
   BadgeStatutAction,
-  Box,
   Card,
   Paragraph,
   Stack,
@@ -17,68 +16,7 @@ type ActionLieeCardProps = {
 };
 
 const ActionLieeCard = ({ action }: ActionLieeCardProps) => {
-  const { identifiant, nom, referentiel, statut, actionId } = action;
-
-  const pointsFait = action.score.pointFait;
-  const pointsPotentiel = action.score.pointPotentiel;
-
-  const troncateIfZero = (value: string | undefined) => {
-    if (!value) {
-      return '0';
-    }
-    return value.endsWith('.0') ? value.slice(0, -2) : value;
-  };
-
-  const calculatePercentage = (
-    pointFait: number | null,
-    pointPotentiel: number | null
-  ): string => {
-    if (pointFait && pointPotentiel) {
-      const percentage = ((pointFait / pointPotentiel) * 100).toFixed(1);
-      return troncateIfZero(percentage);
-    }
-    return '0';
-  };
-
-  const roundPointFait = troncateIfZero(pointsFait?.toFixed(1));
-  const roundPointPotentiel = troncateIfZero(pointsPotentiel?.toFixed(1));
-
-  // Barre de score
-
-  const score = [
-    {
-      label: avancementToLabel.fait,
-      value: action.score.pointFait,
-      color: actionAvancementColors.fait,
-    },
-    {
-      label: avancementToLabel.programme,
-      value: action.score.pointProgramme,
-      color: actionAvancementColors.programme,
-    },
-    {
-      label: avancementToLabel.pas_fait,
-      value: action.score.pointPasFait,
-      color: actionAvancementColors.pas_fait,
-    },
-  ];
-
-  const percentageAgainstTotal = (x: number): number =>
-    (100 * x) / (action.score.pointPotentiel || 1);
-
-  const localData: { label: string; value: number; color: string }[] = [];
-  score.forEach((s, idx) => {
-    if (s.value === 0) {
-      return localData.push({ ...s, value: 0 });
-    }
-
-    const value = s.value ? percentageAgainstTotal(s.value) : 0;
-
-    localData.push({
-      ...s,
-      value: value + (idx >= 1 ? localData[idx - 1].value : 0),
-    });
-  });
+  const { identifiant, nom, referentiel, statut } = action;
 
   return (
     <Card wrap={false} gap={1.5} className="w-[32%] p-3">
@@ -94,49 +32,10 @@ const ActionLieeCard = ({ action }: ActionLieeCardProps) => {
         <Title variant="h6" className="leading-5 text-primary-8">
           {identifiant} {nom}
         </Title>
-
         {/* Badge de score */}
-        <Stack direction="row" gap={0}>
-          {false ? (
-            <Badge title="0 point" state="grey" uppercase={false} />
-          ) : (
-            <>
-              <Badge
-                title={`${calculatePercentage(pointsFait, pointsPotentiel)} %`}
-                state="success"
-                uppercase={false}
-                className="rounded-r-none border-2 border-r-0"
-              />
-              <Badge
-                title={`${roundPointFait} / ${roundPointPotentiel} points`}
-                state="success"
-                light
-                uppercase={false}
-                className="rounded-l-none border-2"
-              />
-            </>
-          )}
-        </Stack>
-
+        <ScoreRatioBadge score={action.score} />
         {/* Barre de progression */}
-        <Stack
-          className={classNames(
-            'relative flex pt-1 min-w-[100px] min-h-[10px] w-full bg-gray-200 rounded-[4px]'
-          )}
-        >
-          {localData
-            .sort((a, b) => b.value - a.value) // Permet d'afficher les plus grandes barres en premières afin d'avoir un effet de stacking et ne pas utiliser de z-index
-            .map((d) => (
-              <Box
-                key={d.label}
-                className={classNames(
-                  'absolute min-h-full top-0 left-0 rounded-[4px]',
-                  `min-w-[${d.value}%]`,
-                  `bg-[${d.color}]`
-                )}
-              />
-            ))}
-        </Stack>
+        <ScoreProgressBar score={action.score} />
       </Stack>
     </Card>
   );
@@ -156,7 +55,7 @@ const ActionsLiees = ({ actionsLiees }: ActionsLieesProps) => {
       </Title>
 
       <Stack gap={3} direction="row" className="flex-wrap">
-        {actionsLiees.map((action: any) => (
+        {actionsLiees.map((action) => (
           <ActionLieeCard key={action.actionId} action={action} />
         ))}
       </Stack>
